Added the service
Refactored data access to the BookService class.
Registered the BookService as the provider of its service at the root level so that it can be injected anywhere in the application.
Angular Dependency Injection to inject it into a component.
HeroService get data method an asynchronous signature.
Discovered Observable and the RxJS Observable library.